final update
had a tutoring session and they helped guide me in the correct approach to take when dealing with this problem. changed all vars to let. changed the logic gate for char length to not            have any else statements. changed the value of characters to a array instead of a string so it can properly index.